Add proof-of-concept to main `name-gen.py`
The included proof-of-concept will generate a specified number of names
using the specification included by the command line. There are numerous
points of improvement here: modulization, input checking, etc., and
probably an easier way to use it than the command line entry.

The proof-of-concept demonstrates two things: (1) the Zipfian picker
works as expected and (2) the program can correctly read a specification
file and produce an output.